refactor: move primary templates and helper alias/constants for all traits into Fwd.hpp
